refactor: package ian application architecture
