+++ .gitignore
@@ -35,5 +35,4 @@ Thumbs.db
 # Project specific
 uploads/
 data/
-=3.10
-25.1.1
+*.egg-info/
