+++ src/ian/domain/__init__.py
@@ -1,0 +1,2 @@
+"""Pure domain logic for NTUAI Agent."""
+
