+++ src/ian/gateways/__init__.py
@@ -1,0 +1,2 @@
+"""Platform gateway adapters."""
+
