+++ src/ian/services/__init__.py
@@ -1,0 +1,2 @@
+"""Stateful and I/O-backed services."""
+
