+++ tests/agent/test_agent_placeholder.py
@@ -1,0 +1,6 @@
+import pytest
+
+pytest.skip(
+    "Agent/MCP/LLM integration tests are intentionally left as placeholders.",
+    allow_module_level=True,
+)
